Added default workspace/exposure location to the global settings object, making way to remove the magic guessing of workspace and exposure object locations.
Also, added basic support for storing of absolute location of the source
workspace for an exposure by supporting the resolution to workspace from
this value on top of the old method via the pattern in the stored value.
This change also enable resolution of workspaces within user workspaces,
which have uri in the form `${user}/${workspace_id}`.

Further testing will be needed once user workspaces are implemented.

Progress towards:
https://tracker.physiomeproject.org/show_bug.cgi?id=2368

git-svn-id: https://svn.physiomeproject.org/svn/physiome/pmr2.app/trunk@4581 eb92df5b-a40c-0410-b02c-bf52feaaa166